More resilient error parsing
If we run this script with -x, then the commands are echoed as they are
executed.  This will echo the awk line, which means awk will find its
own command line in the log file, and print the bad line twice, which
will make the JSON unparsable.

By changing the pattern to /to +retry,/ we still find the log line we
want, but the pattern doesn't match itself, so it still works when run
with -x.